+++ src/main/kotlin/party/para/model/JoinChaperoneRequest.kt
@@ -1,10 +1,12 @@
 package party.para.model
 
+import java.math.BigDecimal
+
 data class JoinChaperoneRequest(
     var hospital : Int,
     var name: String,
     var startHour : Int,
     var endHour : Int,
-    var price : Float,
+    var price : BigDecimal,
     var phone : String
 )
